deploy to aws lambda (#1)
I converted the app to Flask to make it easier to deploy to lambda. I figure async probably isn't super necessary on lambda anyway since AWS creates an app per request.

I'm not happy with Zappa after a very glitchy start to setting up the lambda app, but I think it's okay for now. Zappa is a wrapper around lambda for Python apps.